Standardize naming in ParticleCircle
This is a follow-up to https://github.com/GitBrincie212/Apel-Mod/pull/1
to standardize on `beforeDraw`, `afterDraw`, `drawPos`, and
`particleEffect`.  It also passes specific named objects of the current
type to `doBeforeDraw` and `doAfterDraw` for increased clarity. Field
visibility of `DrawInterceptor` implementations is adjusted and set via
setter to ensure the correct interceptor is present and called, if
present.

This also telescopes the constructor so logic isn't repeated in
constructors.  This follows a best practice to have one constructor and
one copy constructor per class.

The interceptors now have setter methods, and the radius has
getter/setter.  Therefore, document those methods so others know how to
use them.