Standardize ParticleCuboid naming and structure
This makes the same changes to `beforeDraw`, `afterDraw`,
`particleEffect`, and `drawPos` as done to other `ParticleObject`
subclasses.

Additionally, add defensive copies when constructing the cuboid so the
caller can't inadvertently change the cuboid if they reuse the
`Vector3f` instance.

Extract a `getSize()` method because both `setSize` methods need it, as
do the new unit tests added to ensure that setting the size via `float`
or via `Vector3f` have the same result.