Make ParticleObject abstract
ParticleObject contained `setBeforeDraw` and `setAfterDraw` methods
because it was serving two purposes: an abstract base and a particle
point. By making it abstract, it no longer has to serve as a particle
point class. Therefore, extract a proper `ParticlePoint` class, which
also does a decent job at demonstrating the intended structure for a
`ParticleObject` subclass.

Recognize that `endDraw` isn't used, and remove it for now.  There may
be a future where `beginDraw`, `endDraw`, and `draw` are all provided,
but for now, the subclasses are handling this via their `beforeDraw` and
`afterDraw` `DrawInterceptor`s.