Standardize ParticleQuad
This is a follow-up to https://github.com/GitBrincie212/Apel-Mod/pull/1
to standardize on `beforeDraw`, `afterDraw`, `drawPos`, and
`particleEffect`.  Field visibility of `DrawInterceptor` implementations
is adjusted and set via setter to ensure the correct interceptor is
present and called, if present.

This also telescopes the constructor so logic isn't repeated in
constructors.  This follows a best practice to have one constructor and
one copy constructor per class.  For `ParticleQuad`, this means a
private static method computes the rectangle (or square) vertices.

The interceptors now have setter methods; therefore, document those
methods so others know how to use them.

Also, fix a rotation issue (rotateY took the x-value).