Standardize ParticleSphere
This is a follow-up to https://github.com/GitBrincie212/Apel-Mod/pull/1
to standardize on beforeDraw, afterDraw, drawPos, and particleEffect.
Field visibility of DrawInterceptor implementations is adjusted and set
via setter to ensure the correct interceptor is present and called, if
present.

This also telescopes the constructor so logic isn't repeated in
constructors. This follows a best practice to have one constructor and
one copy constructor per class.

Fixes the offset logic, which was applying prior to rotating the sphere,
and thus was bounded by the trig functions involved in pitch, yaw, and
roll.

Also moves the after-draw interceptor to be after drawing each particle
instead of before drawing.

Documents and defaults the interceptors.