Remove `InterceptorResult`
This is a step toward simplifying interceptors and preventing unintended
consequences.  Since the user should not render a different object than
the current one being intercepted, do not allow returning a reference to
a `ParticleObject`.  Also, do not allow substituting a different
`InterceptData` instance.  With both return values removed, the
`DrawInterceptor` no longer has a value to return and becomes a void
function.  The identity concept is still there and used to ensure
correctness.

`InterceptData` has a bunch of new documentation and one important new
method: `getMetadata(key, defaultValue)`. It is common in Java now to
have maps with default values, and this default value can provide a
useful piece of information: the desired return type.  With this, Apel
can cast the return value for the user and apply a default value to
avoid unboxing exceptions.  The onus is still on the user to provide a
useful default value.