Allow appending additional objects to ParticleCombiner
Using `Arrays.asList` (and similar methods) often results in
unmodifiable lists.  This means future calls to `setObject`,
`setOffset`, `appendObject`, and `appendObjects` will not work.  Handle
this by creating `ArrayList`s when needed.

Also fix a case where the same reference to an offset was reused.  If
the exact same instance of the offset vector is required, the developer
should explicitly set it via multiple calls to `setOffset(index,
offset)` or set it directly on the child objects.