Abstract ServerWorld behind ApelRenderer
By inserting a renderer, the developer can now control how and where
particles are rendered.  Right now, the only option is still via
`ServerWorld`, so this seems simple.  Future options will include the
ability to "bake" an animation, send the rendering instructions to the
client, and potentially more destinations.

`ApelRenderer` is an interface that all renderers should satisfy, and
it is currently used as a factory for producing instances.  This may
change, but that will evolve as more renderers come along and have one
or more options to configure them.

Once an animation is baked (perhaps via `ApelBakingRenderer`), loading
that file to a (future) `ReplayAnimator` would then be able to use the
existing `ApelServerWorldRenderer` to display it.  Once packets are
ready, it could use a `ApelPacketRenderer`.