Send a payload encoding the current frame
The payload encodes the current frame into a list of `Instruction`
objects which are implemented as "record classes" (new with Java 16) and
leverages a switch expression that is confined to only `Instruction`
types (Java 21).

The `ApelNetworkRenderer` will encode these instructions, and the
`ApelFramePayload` will serialize and deserialize using MC's `Packet`
and `RegistryByteBuf` concepts.  Conceivably, the `Instruction` classes
could (should?) be moved into the `ApelRenderer` interface and used from
the `ParticleObject` hierarchy.  Then both the default and network
renderers become a little simpler.

A future `ApelFileRenderer` (for pre-computing animations) will use this
same `Instruction` set.