Cache the sphere points
Prior to this commit, the cache held a single point for a single index,
but every time the `computeCoords` function was called, it was called
with a different index, so the cache didn't accomplish anything.

Now, cache all coordinates from the "golden spiral" method, and
recompute only when the `amount` field changes.  Note that the copy
constructor does _not_ make copy of the vector, and that's acceptable
because the distribution is deterministic for a given `amount`.  When
the particles are drawn, they are first rotated, and that generates a
new `Vector3f` object instead of modifying the cached vector objects.