Rotate sphere points like all others
Sphere had its own rotation logic that computed the cumulative Z, Y, and
X-axis rotations prior to drawing each point on the sphere.  While this
is useful to cache and apply in a single function instead of three, the
rotation logic and caching should be done more centrally, possibly even
via computed quaternions, which should be faster than Euler angles.