Fix Iterator concurrent modification bug
The `ApelScheduler`'s `scheduledTasks` is iterated in the `runTick`
method.  The `Iterator` created there is allowed to add and remove
elements from the list, and it removes a sequence when one finishes.
However, due to using an Iterator here, if another event in the
Minecraft server causes a new `ScheduledSequence` allocation to occur
while `runTick` is iterating, a `ConcurrentModificationException` will
be thrown.

This fix iterates by accessing elements directly through a for-loop
index, which already happened with the companion `animators` list.  It
also takes the opportunity to do a better job of encapsulating logic and
fields in the classes where they should be.