Add Bezier instruction
Uses 'B' as the code, followed by the number of control points.  It then
proceeds to read/write the position, start, control points, end,
rotation, and particle count for each Bezier curve.  One instruction
corresponds to one curve, so if a `ParticleBezierCurve` contains `N`
curves, then `N` `Instructions` will be sent.

This also adds a mechanism by which the control points can be
retrieved and properly scopes the `start` and `end` points to be
`protected`.