Correct arccosine domain and better encapsulate TrigTable
The trig table for arccosine values is computing the indices for
arccosine incorrectly.  This results in ArrayIndexOutOfBoundsException
while rendering objects like spheres that need the full domain of
arccosine.  Arccosine has a domain in the interval [-1, 1]
and a range in the interval [0, PI]. Therefore, compute a different
interval for stepping through the arccosine function and a different
normalization function when looking up an arccosine value.

Encapsulation is improved by generating the trig tables upon creation
rather than requiring a second interaction with the class to initialize
fully.

Also, add documentation for public methods.