Cache Instruction points on the client renderer
Build an Instruction -> Vector3f[] cache for use on the client side.
Instructions should be used because they're `record` classes that
auto-generate equals and hashcode implementations, which makes them
perfect for use as `Map` keys.  The resulting `Vector3f[]` must be
copied prior to use, but it will save N operations per point per shape
if the only operations needed prior to rotation are lookup, copy
constructors for `Vector3f`, and `Quaternionf` construction.

It is possible that a Quaternion cache may be useful, but it is common
for the rotation to update on each server tick.

It is also likely that this caching mechanism can be used by the
`DefaultApelRenderer` since it performs similar caching to what was
originally attempted in the `ParticleObject` subclasses that required
significant trigonometry.

Overall, this is an optimization, but a lesser one than it would have
been prior to the `TrigTable`'s introduction.  Further optimization on
the client side will come when I can reuse `Particle` objects within the
`ParticleManager` and/or render our own particles such that even the
quad used for rendering can be saved/updated instead of recomputed every
single time.