Refactor BezierCurve Instruction
The `Instruction` for Bezier curves was detailed, component-wise, and
this isn't actually necessary or useful.  As the client rendering is
being written, it's clear that it would be simpler to work with a curve
object instead of the three constituent data members.  The abstract
methods on the `BezierCurve` along with the accessors for start, end,
and control points are sufficient.

Therefore, introduce a static factory method for use by the
`BezierCurve` `Instruction` when sending to the client and receiving in
the client.

It seems possible, from this refactoring and the previous one for
ellipses, that the server rendering can also make use of these
`Instruction` objects without loss of performance, especially if the
point caching on `Instruction` objects plays out as pictured.  In short,
an `Instruction` knows everything necessary to compute its original
points, so the renderers only need to rotate and offset during each
frame.