Add Instruction for Cone
Move cone drawing into the renderer instead of the `ParticleCone`.  This
`computePoints` is written for the `ParticleManagerRenderer` to use.
is the first object to do a couple new things:

- Bypass `ParticleObject` entirely for drawing.  There's no reason to
call through yet another method just to render a cone when
`ParticleCone` already has the `renderer` reference.  This change will
come to the other objects.

- Cache only the unit object's points.  Since developers can change
height and radius on each frame, it could blow out the cache, and would
require computation on every frame anyway.  This change will come to the
other objects.