Fix ArrayIndexOutOfBoundsException in TrigTable
When testing large numbers of particles on shapes that leverage the
golden spiral, it was possible for them to request the array index just
beyond the array.  For example, the 421st point on the spiral requested
index 700 in a trig table with 700 entries, which only has indices from
0 to 699.  This is because once multipled by `PI` and `sqrt5plus1`, the
resulting value (4285.129257) is almost a multiple of `TAU`.  When
taking the modulo of that number by `TAU`, the result is 6.27993197,
which is closer to `TAU` than the radian value mapped to the actual last
index would be.  To attempt to be as accurate as possible, the computed
index (699.637518) was rounded, which gives 700, and that doesn't exist
in the array.

To fix this, increase the size of the internal arrays by 1.  For `sin`
and `cos`, the final value should be equal to the value in the 0-th
index.  For `arccos`, this just spreads the values across one more
index.  All three normalize indices using `this.size` and rounding, so
the maximum possible index is `this.size` which is one less than the
number of entries in the arrays; therefore, out of bounds exceptions
should no longer occur.