ParametricBezierCurve fixes for compute and length
Parametric suffered the same problem as quadratic and cubic for
`length`, in that it didn't include the ending point.

It also had a problem with computing De Casteljau's algorithm because it
only used the (interior) control points and did not include both the
start and end points.  This code demonstrates the problem:

```
BezierCurve bc = BezierCurve.of(new Vector3f(-2), new Vector3f(2),
                                List.of(new Vector3f(-1), new Vector3f(0), new Vector3f(1)));

assertEquals(new Vector3f(2), bc.compute(1f));  // Fails!  Returns (1, 1, 1)
```

This happened because the `end` point isn't included.  Correct this by
constructing an internal list that includes all of start, control, and
end points.

Tests are added to ensure these calculations continue to work.