Ensure rendering arrays are the correct length
Before, an `ArrayIndexOutOfBoundsException` occurred when trying to
access the `renderingSteps` or `renderingInterval` values if the
constructor handling a primitive value was used.

Fixing the array issues resulted in telescoping the constructors so that
only two of the six constructors (excluding the copy constructor) run
actual logic.  The other four delegate with minor argument
transformation, such as converting to array, or filling an array with a
default value.