Untangle `convertToSteps` and `getDistance`
These functions called each other, resulting in a
`StackOverflowException`.

Converting to steps based on an interval does need the distance(s) of
the Bézier curve(s).  However, computing the distance should not be
based on a supposed number of steps.  If steps are provided, use those
to compute the `t`-value when rendering.  If an interval is provided,
compute the length of the Bézier curve by assuming 100 segments on the
curve.  This should be enough for all but the most complicated, twisted
curves.

Once the length of the curve is known, computing the number of steps is
straightforward.  With the number of steps known (either provided or
computed), the `t`-step is simply the inverse of that number.