Demonstrate object builders (Circle and Cone)
This provides the builder pattern for Circle and Cone.  The constructors
become private, and the only method by which objects can be created is
the static `builder()` method.

This does move to use setters everywhere when setting properties from
within a constructor to reuse validity checks in both constructors and
setters.

With the move to reuse setters to validate constructor parameters, it is
no longer safe to allow overriding setters.  The setter in a subclass
that overrides one of ParticleObject's setters could be called before
the superclass (ParticleObject) is done initializing, and that can lead
to the subclass depending on its own logic that has not yet run.  To
guard against this, all setters and builder methods in ParticleObject
are final.

It should be considered whether to do this for the set of subclasses.

This does lead to some name changes in classes that have multiple
amounts like BezierCurve, Combiner, and Cuboid since they can no longer
 override `setAmount`, `setRotation`, `setOffset`, or
 `setParticleEffect`.