Use type-safe heterogeneous container for DrawContext
Prior to this, every class had to define enums for keys, and the types
were not enforced at compile time.

A type-safe heterogeneous container (THC) allows class-based keys, and
accessor methods can leverage the generic type to ensure that both
insertion and retrieval are type-safe at compile time.  `DrawContext`
uses the `TypeToken` from Guava that knows how to extract the
superclass' generic type parameter.  This does require that the class be
explicitly declared; a generic method will result in the generic type
being 'T', 'R', etc., as seen in the source code.

This particular implementation performs equals and hashCode on the key
type and the key name, but NOT the actual class.  Equals generally uses
the value of `getClass()`, but if multiple keys are defined that share
a name and a type, they must refer to the same entry in the
`DrawContext`'s metadata map.