Use the `DrawContext` when intercepting objects
ParticleObject now holds the interceptors, so it must be
generic to provide correctly typed parameters to subclasses when
executing interceptors.

ParticleObject also provides a new method for animators to call:
`doDraw`.  This retains the same call signature, as there is no need to
have Animators understand a `DrawContext` since they don't draw.  The
`doDraw` method asks the subclass to `prepareContext`; this is an
opportunity for the subclass to add metadata values prior to the
`beforeDraw` interceptor.  Often, this will provide some pre-computed
values that will be used in `draw` so that interceptors can modify them.
Since the `draw` method can easily retrieve these values, this works
well to allow modification prior to drawing.  It then calls the
interceptor, the draw method, and the afterDraw interceptor.  Subclasses
that want to provide values to `afterDraw` may do so as part of their
`draw` implementation.

ParticleBezierCurve has a somewhat unique `draw` implementation because
it draws multiple objects (similar to ParticleCombiner).  As a result,
it needs to manipulate the `drawContext` inside its for-loop and then
call a pair of interceptors around drawing each child.  The assumption
right now (perhaps flawed?) is that the same interceptor that one would
call on a singular object is what is desired around each Bézier curve
object.  If this assumption does not hold, then ParticleBezierCurve will
either mimic ParticleCombiner and host its own interceptors, or it will
switch to drawing a single curve, and those that wish to draw multiple
curves will construct a ParticleCombiner.  The latter moves toward a
simpler, more consistent set of functionality.

ParticleCombiner is no longer generic because typing through static
builders generally requires the type parameter to be specified (e.g.,
ParticleCombiner.<ParticlePoint>builder()), and adding objects of
multiple types requires a `ParticleObject` type parameter.  Since this
is the standard type that most would see, go ahead and remove it.

ParticleCombiner also doesn't have `beforeDraw` and `afterDraw`;
instead, it has its own `beforeChildDraw` and `afterChildDraw`
interceptors that are still configured via builder, but are not passed
to the superclass.