Use `doDraw` when drawing Combiner children
A combiner with a cuboid and a circle (with the cuboid added first) was
not drawing anything when activated.  This is because Cuboid uses
`prepareContext` to set the `VERTICES` metadata so an interceptor can
modify the vertices (if desired).  With Combiner, it wasn't calling
`doDraw`, which includes the `prepareContext` call.  Therefore, have
Combiner call `doDraw`, which gives each of its children the expected
sequence during the drawing step.