Render all frames when `processSpeed` > 1
The `handleDrawingStep` method was skipping the frame that occurred when
scheduling the step.  It collected all frames when a step wasn't ready
to be scheduled (when `step % processSpeed != 0`).  This would drop one
out of N frames, which is incorrect.

Now, if delay != 0 and processSpeed > 1, always add the runnable to the
list.  Then, when the list is long enough, schedule the step with the
appropriate delay and clear the list.