Convert PointAnimator to builder
This is the simplest animator, so the builder is very straightforward
to implement.  Rather than overload the term 'origin', change to
'point'.  Getter and setter are available for point, though if one
is going to use these, perhaps another animator is ideal.

No validity checks on the provided point at this point, though perhaps null should be checked...