Convert Circular/EllipseAnimators to builder
Straightforward addition of builders.  Should consider which properties
are changeable after building.  Center and radius should be OK, rotation
is likely OK.  Revolutions should not be, as it is used to compute the
number of steps before the animation starts.  Clockwise can be toggled
without issue.  Trimming is not used, but would be OK if it changed (at
this point).