Convert BezierCurveAnimator to builder
This switches the internal tracking from arrays to lists.  This does
consume marginally more memory (overhead for the `List` object), but
will work better for a builder, as it's much easier to append to lists
than arrays.

The Builder allows for adding steps per curve, interval per curve, steps
for all curves, or an interval for all curves.  The priority is given to
intervals over steps, and also to the "allCurves" variant vs. the
per-curve variant.

The BezierCurveAnimator does not currently respect the trimming value.
It should respect it, but it should be a float-based trim on the value
of `t` as the curve is drawn.  It would be possible to trim in multiple
ways if multiple trim objects were allowed.