Change Cuboid handling of 'amounts'
The cuboid previously separated the top face, bottom face, and vertical bars into
separate amounts.  This was always awkward, since the size of the cube isn't set
this way.  It is likely that most cuboids will want similar numbers of particles
on edges in a given direction, so switch the amount handling to follow the same
practice as the edge length.  This means that all edges parallel to the x-axis
will have the same number of particles rendered, and the same is true for y and z.