Provide a basic ParticleArray
This will allow repeating a single object in x, y, and z directions with
configurable number and spacing of the repeated object.  This is a very
basic starting point from which more features may be added, such as
rotating and/or offseting the entire array, intercepting between
individual renders, and more.

Note that the builder is a little different for this one.  To get the
compiler to understand the multiple recursive generic parameters, the
particle object must be provided when constructing the builder instead
of (or in addition to) the usual builder approach.  There may be a way
to resolve this, but I did not figure it out if it exists, yet.

`particleEffect` and `amount` are disregarded, even though they're
passed to the superclass constructor.

Test code:
```
private void rotator(DrawContext data, ParticleObject<? extends ParticleObject<?>> obj) {
    obj.setRotation(obj.getRotation().add(0f, 0.01f, 0f));
}

private void arrayTest(ServerWorld world, PlayerEntity user) {
    ParticleCircle circle = ParticleCircle.builder()
            .particleEffect(ParticleTypes.BUBBLE)
            .radius(1f)
            .amount(20)
            .afterDraw(this::rotator)
            .build();

    ParticleArray<ParticleCircle> array = ParticleArray.builder(circle)
            .gridSize(new Vector3i(2, 3, 4))
            .spacing(new Vector3f(2f, 3f, 4f))
            .build();

    PointAnimator pointAnimator = PointAnimator.builder().point(new Vector3f(0, 5,0)).delay(1).renderingSteps(600)
            .particleObject(array).build();
    pointAnimator.beginAnimation(ApelServerRenderer.client(world));
}
```