Document and clean up Linear/Bezier animators
The documentation was lacking on these animators to help understand
how they construct themselves and the relationship between steps and
intervals.

Also, realize that these two animators do not use the base class' interval
nor steps, so the checks being done for those should not occur unless they
are set.  Use the enum to determine whether the setters are necessary or
not.  Do this in all animators (parallel and sequential did not need
changes).

AnimationTrimming also receives a copy constructor for use when building
animators so that once checked, callers can no longer modify the trim
values prior to building the animator instance.  This does not handle
the getter/setter on the animator itself.