Create an AnimationContext
`AnimationContext` mimics `DrawContext`, but with properties specific to
`Animator` implementations.  Tests check that `AnimationContext`
functions in the same way as `DrawContext`.

This sets the stage to extract the `Key` class from `DrawContext` since
it's used by both contexts.  The pre-defined `Key` implementations will
move to either the `Key` class or a separate `Keys` class, if needed.