Use AnimatorInterceptor in all animators
The new `AnimationInterceptor` mimics the `ObjectInterceptor`, yet
allows for different included properties.  The context provided allows
additional metadata that functions exactly as those added to
`DrawContext`.  It replaces the (renamed) `OldInterceptors` approach
that used the previous interceptor implementation.

Introducing this requires parameterizing `PathAnimatorBase` to ensure
the type of the `PathAnimatorBase` provided to the interceptor matches
the animator type.  This allows access to properties of the specific
animator subclass, not just those of the base class.

The current step is not provided, currently, though it should be trivial
to add.  There are potentially some other properties (e.g., step per segment
in `LinearAnimator`) that could be useful as well.  Those will come in
separate commits to demonstrate clearly how to add new properties.