Add `currentStep` to the `AnimationContext`
`AnimationContext` needs to be able to provide the step number to the
interceptor so that every individual animator doesn't have to define or
reference a `Key` and provide it via metadata.  Since it's a common
property to most animators (excepting Parallel and Sequential, which are
utility animators), it makes sense to include here.

Note: this commit was deliberately separate from introducing the
`AnimationContext` for the purpose of demonstrating how to add a single
parameter to the `AnimationContext`.